work. add realtime update parameters

+++ app/src/main/java/com/lk/biocadproject/models/ParametersModel.kt
@@ -1,10 +1,10 @@
 package com.lk.biocadproject.models
 
-data class ParametersModel (var pressure:Float,
-                            var humidity:Float,     // влажность измеряется от 0 до 100%
-                            var roomTemperature:Float,
-                            var workingAreaTemperature:Float,
-                            var levelPH: Float,
-                            var weight: Float,
-                            var fluidFlow: Float,
-                            var levelCO2: Float)
+data class ParametersModel (var pressure:Double,
+                            var humidity:Double,     // влажность измеряется от 0 до 100%
+                            var roomTemperature:Double,
+                            var workingAreaTemperature:Double,
+                            var levelPH: Double,
+                            var weight: Double,
+                            var fluidFlow: Double,
+                            var levelCO2: Double)
